First implementation of rust-gpx
This commit contains the first implementation of rust-gpx. It's a rough
implementation, but it does a chunk of the major features, enough to
entirely parse the GPX example from Wikipedia. Currently, only the
parser is implemented, and it does not support emitting new GPX files.

The parser uses xml-rs, a pretty simple XML parser. Alternatives can be
explored in the future (it doesn't seem to be the fastest choice), but
at least for the near future, we'll be continuing with xml-rs.

As of right now, there's relatively minimal interaction between gpx and
geo. This will change in the future as we add APIs that allow GPX data
to actually be consumed in a meaningful way, as opposed to just loading
the data straight from the file. This will aid in adoption with GPS
users.

The library still needs a few changes until we can do our first release,
but we're definitely close.

This contribution is licensed under the MIT license, (c) Brendan
Ashworth.