release 0.1.0
This commit is the first real release of rust-gpx. It includes basic
read support for a lot of GPX use cases, although it does not have
comprehensive support of the GPX spec. At this time it can be
effectively used for decoding many GPX files however, especially
activity files created by GPS tools such as watches.